% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{agenda.add}
\alias{agenda.add}
\title{Little helper adds an agenda item to the agenda dataframe}
\usage{
agenda.add(text, start, end, comments = "", data = agenda)
}
\arguments{
\item{text}{a description of the agenda item}

\item{start}{the start level of the mon you're (presumably) leveling}

\item{end}{the level you're trying to reach}

\item{comments}{comments}

\item{data}{the name of the agenda table you're adding to}
}
\description{
Little helper adds an agenda item to the agenda dataframe
}
