% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{alldps}
\alias{alldps}
\title{Lists all combinations of moves for a given 'mon, and their
average combined DPS per totaldps(). As with that, can adjust
the numbers to show how they perform against a target 'mon.
Ends with ratio of damage to specific target to unadjusted DPS.}
\usage{
alldps(mon, target = NA, inc.old = FALSE, inc.rank = TRUE)
}
\arguments{
\item{mon}{The pokemon in question}

\item{inc.old}{Show movesets no longer seen in the wild}

\item{inc.rank}{Exclude the offense/defense ranks}
}
\description{
Lists all combinations of moves for a given 'mon, and their
average combined DPS per totaldps(). As with that, can adjust
the numbers to show how they perform against a target 'mon.
Ends with ratio of damage to specific target to unadjusted DPS.
}
