% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{allmoves.add}
\alias{allmoves.add}
\title{Add a new move to allmoves}
\usage{
allmoves.add(name, type, speed, power, seconds, energy, notes = NA)
}
\arguments{
\item{name}{the new move's name}

\item{type}{the move's type}

\item{speed}{the move's speed (Fast or Charge)}

\item{power}{the move's power}

\item{seconds}{the move's duration in seconds}

\item{energy}{the move's energy generation/cost}

\item{note}{any special notes for the move (e.g. misc. effect)}
}
\description{
Add a new move to allmoves
}
