% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{combos.better}
\alias{combos.better}
\title{Counts the number of combinations that are better than one with given stats.}
\usage{
combos.better(
  mon,
  atk = 10,
  def = 10,
  sta = 10,
  minstr = 10,
  mindef = 10,
  minsta = 10
)
}
\arguments{
\item{mon}{what kind of mon you're looking at}

\item{atk}{your mon's attack IV}

\item{def}{your mon's defense IV}

\item{sta}{your mon's stamina IV}

\item{minstr}{the minimum attack IV for result combos}

\item{mindef}{the minimum defense IV for result combos}

\item{minsta}{the minimum stamina IV for result combos}
}
\description{
Counts the number of combinations that are better than one with given stats.
}
