% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{combos}
\alias{combos}
\title{Counts the number of combinations that equal a range, with a min roll of min and max max}
\usage{
combos(equals = 0:45, min = 0, max = 15, print = FALSE)
}
\arguments{
\item{equals}{the range of sum of stats you want to check, from 0 to 45}

\item{min}{the minimum score for each stat}

\item{max}{the maximum score for each stat}

\item{print}{print each combination or no?}
}
\description{
Counts the number of combinations that equal a range, with a min roll of min and max max
}
