% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{copy_to_clipboard}
\alias{copy_to_clipboard}
\title{Copy info to the clipboard (Mac version)}
\usage{
copy_to_clipboard(x, sep = "\\t", col.names = T, ...)
}
\arguments{
\item{x}{The data to copy to the clipboard}

\item{sep}{The cell separator}

\item{col.names}{First row has column names}
}
\description{
Copy info to the clipboard (Mac version)
}
