% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{cpcalc}
\alias{cpcalc}
\title{Calculates CP for a given mon using level, attack, defense, and stamina IVs.
Defaults to calculating max CP.}
\usage{
cpcalc(mon, level = 40, attack = 15, defense = 15, stamina = 15)
}
\arguments{
\item{mon}{The pokemon}

\item{level}{Level, default 40}

\item{attack}{Attack, default 15}

\item{defense}{Defense, default 15}

\item{stamina}{Stamina, default 15}
}
\description{
Calculates CP for a given mon using level, attack, defense, and stamina IVs.
Defaults to calculating max CP.
}
