% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{CPin}
\alias{CPin}
\title{Calculate CP for every mon in your database}
\usage{
CPin(data = mymons)
}
\arguments{
\item{data}{the name of your database}
}
\description{
Calculate CP for every mon in your database
}
