% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{hpcalc}
\alias{hpcalc}
\title{Calculates HP for a given mon using level and stamina IV. Defaults to calculating max HP.}
\usage{
hpcalc(mon, level = 40, stamina = 15)
}
\arguments{
\item{mon}{The pokemon}

\item{level}{Level, default 40}

\item{stamina}{Stamina, default 15}
}
\description{
Calculates HP for a given mon using level and stamina IV. Defaults to calculating max HP.
}
