% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{HPin}
\alias{HPin}
\title{Calculate HP for every mon in your database}
\usage{
HPin(data = mymons)
}
\arguments{
\item{data}{the name of your database}
}
\description{
Calculate HP for every mon in your database
}
