% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{levelcalc}
\alias{levelcalc}
\title{Calculates the level corresponding to a given bunch of stats.}
\usage{
levelcalc(name, CP, HP, Dust, Attack, Defense, Stamina)
}
\arguments{
\item{name}{The pokemon}

\item{CP}{its CP}

\item{HP}{its HP}

\item{Dust}{dust to next level}

\item{Attack}{its attack stat}

\item{Defense}{its defense stat}

\item{Stamina}{its stamina stat}
}
\description{
Calculates the level corresponding to a given bunch of stats.
}
