% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{levfromstats}
\alias{levfromstats}
\title{Calculates level and HP given a name, CP, attack, defense, and stamina (convenience for mylist.new)}
\usage{
levfromstats(name, cp, atk, def, sta)
}
\arguments{
\item{name}{the mon's type}

\item{cp}{its CP}

\item{atk}{its attack}

\item{def}{its defense}

\item{sta}{its stamina}
}
\value{
the level number
}
\description{
Calculates level and HP given a name, CP, attack, defense, and stamina (convenience for mylist.new)
}
