% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{loadmonlist}
\alias{loadmonlist}
\title{Loads the database of your pokemon. Careful not to erase important changes! Save early, save often}
\usage{
loadmonlist(data = mymons)
}
\arguments{
\item{data}{Which data file are you saving to?}
}
\description{
Loads the database of your pokemon. Careful not to erase important changes! Save early, save often
}
