% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{loadpoke}
\alias{loadpoke}
\title{Loads all the basic info about pokemon, their moves, and game constants}
\usage{
loadpoke()
}
\description{
Loads all the basic info about pokemon, their moves, and game constants
}
