% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{matcheck}
\alias{matcheck}
\title{Find what level you can afford to get to with specified dust and/or candy}
\usage{
matcheck(level = 1, dust = 1e+07, candy = 10000)
}
\arguments{
\item{level}{your mon's starting level}

\item{dust}{how much dust you have (default=ten million AKA ~unlimited)}

\item{candy}{how much candy you have (default=ten thousand AKA ~unlimited)}
}
\value{
a couple numbers
}
\description{
Find what level you can afford to get to with specified dust and/or candy
}
