% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{matcost}
\alias{matcost}
\title{Either give a starting level or a starting amount of dust,
get how many mats it takes to reach the 'stop' level.}
\usage{
matcost(start = 1, stop = 40, dust = NA)
}
\arguments{
\item{start}{The starting level}

\item{stop}{The level you're trying to get to}

\item{dust}{Provide dust amount to reach next level if you're unsure about which level you're starting at}
}
\value{
The amount of dust and candy required to reach the level
}
\description{
Either give a starting level or a starting amount of dust,
get how many mats it takes to reach the 'stop' level.
}
