% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{moncat}
\alias{moncat}
\title{Lists all pokemon in the database matching a grep for the given string}
\usage{
moncat(name, target = "", order = NA, data = mymons)
}
\arguments{
\item{name}{The name or name fragment you're looking for}

\item{target}{A target to calculate DPS against?}

\item{order}{The order in which to display results}

\item{data}{The database}
}
\description{
Lists all pokemon in the database matching a grep for the given string
}
