% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{moncat.types}
\alias{moncat.types}
\title{Lists all your mons of a particular type}
\usage{
moncat.types(type, data = mymons)
}
\arguments{
\item{type}{the type of mon}

\item{data}{name of your database}
}
\description{
Lists all your mons of a particular type
}
