% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{monsOfType}
\alias{monsOfType}
\title{Looks up all pokémon who are of a particular type. Also shows max CP, to help pick one to pursue.}
\usage{
monsOfType(type, decreasing = F)
}
\arguments{
\item{type}{The type}

\item{decreasing}{In decreasing order}
}
\description{
Looks up all pokémon who are of a particular type. Also shows max CP, to help pick one to pursue.
}
