% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{moveset}
\alias{moveset}
\title{Pass it a pokemon name ("Charizard") and it displays a little info about each move
in the pokemon's possible movesets.}
\usage{
moveset(mon, inc.old = FALSE, bydef = FALSE)
}
\arguments{
\item{mon}{The pokemon name to look up}

\item{inc.old}{Include moves no longer encountered in the wild (e.g. Community Day, obsolete)}

\item{bydef}{Order by defense rank}
}
\value{
Display of move combinations
}
\description{
Pass it a pokemon name ("Charizard") and it displays a little info about each move
in the pokemon's possible movesets.
}
