% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.add}
\alias{mylist.add}
\title{Adds a pokemon to the database conforming to the output of possibles()}
\usage{
mylist.add(poss, fast.move = "", charge.move = "", notes = "",
  shiny = FALSE, stats = TRUE, data = mymons)
}
\arguments{
\item{poss}{A pokemon output by possibles() or a facsimile}

\item{fast.move}{The pokemon's fast move}

\item{charge.move}{The pokemon's charge move}

\item{notes}{Anything for the optional notes field}

\item{shiny}{Is the mon shiny?}

\item{stats}{Are we including stats or not?}

\item{data}{Which database file?}
}
\value{
The tail of the revised database
}
\description{
Adds a pokemon to the database conforming to the output of possibles()
}
