% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.counter}
\alias{mylist.counter}
\title{Shows how much damage a mon in your database could do against a target
This is called by most of the other 'find damage against x' functions like mylist.raidcounters()}
\usage{
mylist.counter(
  sourceID,
  target,
  raid = F,
  targetdef = 15,
  targetlvl = 40,
  Atk = NA,
  Level = NA,
  Fast = NA,
  Charge = NA,
  data = mymons,
  exclude = NA,
  ...
)
}
\arguments{
\item{sourceID}{the row number in your database of the attacker}

\item{target}{the name of the kind of pokemon you're attacking}

\item{raid}{if the target is a raid boss}

\item{targetdef}{the target's defense IV}

\item{targetlvl}{the target's level}

\item{Atk}{specify a different attack IV for your attacker}

\item{Level}{specify a different level for your attacker}

\item{Fast}{specify a different fast attack for your attacker}

\item{Charge}{specify a different charge attack for your attacker}

\item{data}{name of your database}

\item{exclude}{exclude a certain type of mon, I guess?}
}
\value{
one particularly formatted line summarizing the matchup
}
\description{
Shows how much damage a mon in your database could do against a target
This is called by most of the other 'find damage against x' functions like mylist.raidcounters()
}
