% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.counters}
\alias{mylist.counters}
\title{Calls mylist.counter() for each of your mons, defaults to not a raid boss}
\usage{
mylist.counters(target, raid = F, numreturn = 10, Name = "",
  data = mymons, ...)
}
\arguments{
\item{target}{the kind of mon you're attacking}

\item{raid}{is it a raid boss?}

\item{numreturn}{the number of results to print}

\item{Name}{the pokemon kind to limit results to}

\item{data}{the name of your database}
}
\description{
Calls mylist.counter() for each of your mons, defaults to not a raid boss
}
