% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.dpslvl}
\alias{mylist.dpslvl}
\title{Sees how effective one of your mons is at attacking a type of target at each of
a range of your mon's levels. If range is not included, will look for the mon's
level and go from there to 40. If the mon has no level, will go from 20:40.}
\usage{
mylist.dpslvl(sourceID, target, raid = T, range = NA, data = mymons,
  ...)
}
\arguments{
\item{sourceID}{the row number of your attacking mon}

\item{target}{the type of target you're going after}

\item{raid}{is the target a raid boss?}

\item{range}{number range of levels you're checking, e.g. 32:40}

\item{data}{the name of your database}
}
\description{
Sees how effective one of your mons is at attacking a type of target at each of
a range of your mon's levels. If range is not included, will look for the mon's
level and go from there to 40. If the mon has no level, will go from 20:40.
}
