% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.evolve}
\alias{mylist.evolve}
\title{Use to evolve a pokemon in your database, calculating new CP/HP}
\usage{
mylist.evolve(num, to, fast = "", charge = "", cp = NA, hp = NA,
  data = mymons)
}
\arguments{
\item{num}{The mon's row number}

\item{to}{What it's evolving into}

\item{fast}{Optionally, its new fast move}

\item{charge}{Optionally, its new charge move}

\item{cp}{Its new CP}

\item{hp}{Its new HP}

\item{data}{The database you're using}
}
\value{
The mon's old stats and its new stats
}
\description{
Use to evolve a pokemon in your database, calculating new CP/HP
}
