% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.export}
\alias{mylist.export}
\title{Copy a database to the clipboard (Mac)}
\usage{
mylist.export(data = mymons)
}
\arguments{
\item{data}{Name of the database}
}
\description{
Copy a database to the clipboard (Mac)
}
