% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.legendaries}
\alias{mylist.legendaries}
\title{Lists all legendaries in mymons.}
\usage{
mylist.legendaries(data = mymons)
}
\arguments{
\item{data}{the name of your database}
}
\description{
Lists all legendaries in mymons.
}
