% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.level}
\alias{mylist.level}
\title{Change the level of a mon in the list, recalculate CP and HP.
Obviously only works if atk/def/sta are known.}
\usage{
mylist.level(mon, level, data = mymons)
}
\arguments{
\item{mon}{The row number of the mon}

\item{level}{The level it's achieving}

\item{data}{Name of the database}
}
\description{
Change the level of a mon in the list, recalculate CP and HP.
Obviously only works if atk/def/sta are known.
}
