% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.levels}
\alias{mylist.levels}
\title{Attempt to calculate levels of all pokemons in the selected database}
\usage{
mylist.levels(data = mymons)
}
\description{
Attempt to calculate levels of all pokemons in the selected database
}
