% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.new}
\alias{mylist.new}
\title{Add a new 'mon by its stats}
\usage{
mylist.new(
  type,
  cp,
  atk,
  def,
  stam,
  fast = "",
  charge = "",
  charge2 = "",
  shiny = F,
  notes = "",
  data = mymons
)
}
\arguments{
\item{type}{the pokemon type}

\item{cp}{its CP}

\item{atk}{its attack stat}

\item{def}{its defense stat}

\item{stam}{its stamina stat}

\item{fast}{its fast move}

\item{charge}{its charge move}

\item{shiny}{whether it's shiny (default=F)}

\item{notes}{any notes you'd like to add}

\item{data}{what list you're adding to}

\item{chage2}{its second charge move, if any}
}
\description{
Add a new 'mon by its stats
}
