% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.raidcounter}
\alias{mylist.raidcounter}
\title{Wrapper for mylist.counter(), check how effective one of your mons (row number, first
argument) is against a type of raid boss (second argument)}
\usage{
mylist.raidcounter(...)
}
\description{
Wrapper for mylist.counter(), check how effective one of your mons (row number, first
argument) is against a type of raid boss (second argument)
}
