% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.raidcounters}
\alias{mylist.raidcounters}
\title{Wrapper for mylist.counters(), cycles through your mons to see which are best against
a particular target, defaulting to raid versions}
\usage{
mylist.raidcounters(...)
}
\description{
Wrapper for mylist.counters(), cycles through your mons to see which are best against
a particular target, defaulting to raid versions
}
