% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.raidstrong}
\alias{mylist.raidstrong}
\title{Wrapper for mylist.strongvs(), determining which of your mons are strongest against an
attacker, defaulting to raid version}
\usage{
mylist.raidstrong(...)
}
\description{
Wrapper for mylist.strongvs(), determining which of your mons are strongest against an
attacker, defaulting to raid version
}
