% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.statsvs}
\alias{mylist.statsvs}
\title{Shows how strong of both an attacker and defender one of your mons is against a target
across a range of levels -- calls mylist.dpslvl() and mylist.strlvl()}
\usage{
mylist.statsvs(source, target, TarFast, TarCharge, raid = T, range = NA, ...)
}
\arguments{
\item{source}{the row number of your mon}

\item{target}{the kind of attacking mon}

\item{raid}{is the attacker a raid boss?}

\item{range}{the range of your mon's levels to try}

\item{tarFast}{the attacker's fast attack}

\item{tarCharge}{the attacker's charge attack}

\item{data}{the name of your database}
}
\description{
Shows how strong of both an attacker and defender one of your mons is against a target
across a range of levels -- calls mylist.dpslvl() and mylist.strlvl()
}
