% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.strlvl}
\alias{mylist.strlvl}
\title{Shows how strong of a defender one of your mons is against a target across a range of levels}
\usage{
mylist.strlvl(source, target, tarFast, tarCharge, raid = T, range = NA,
  data = mymons, ...)
}
\arguments{
\item{source}{the type of attacker}

\item{target}{the row number of your defender}

\item{tarFast}{the attacker's fast attack}

\item{tarCharge}{the attacker's charge attack}

\item{raid}{is the attacker a raid boss?}

\item{range}{the range of your mon's levels to try}

\item{data}{the name of your database}
}
\description{
Shows how strong of a defender one of your mons is against a target across a range of levels
}
