% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.strong}
\alias{mylist.strong}
\title{See your mon's survivability vs. a certain source's attacks}
\usage{
mylist.strong(
  source,
  target,
  tarFast = "",
  tarCharge = "",
  raid = T,
  sourceatk = 15,
  sourcelvl = 40,
  Def = NA,
  Level = NA,
  data = mymons,
  ...
)
}
\arguments{
\item{source}{the name of the type of attacker}

\item{target}{the row number of your mon}

\item{tarFast}{the attacker's fast attack}

\item{tarCharge}{the attacker's charge attack}

\item{raid}{is the attacker a raid boss?}

\item{sourceatk}{the attacker's attack IV}

\item{sourcelvl}{the attacker's level}

\item{Def}{an alternate for your mon's defense IV}

\item{Level}{an alternate for your mon's level}

\item{data}{the name of your database}
}
\description{
See your mon's survivability vs. a certain source's attacks
}
