% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.strongvs}
\alias{mylist.strongvs}
\title{Iterates mylist.strong() for each mon in your database
That is, calculates the pokemon who take the least damage from a source}
\usage{
mylist.strongvs(source, tarFast = "", tarCharge = "", Name = "",
  raid = T, numreturn = 10, sourceatk = 15, sourcelvl = 40,
  cpmin = 1500, data = mymons)
}
\arguments{
\item{source}{the kind of attacker}

\item{tarFast}{the attacker's fast move}

\item{tarCharge}{the attacker's charge move}

\item{Name}{the kind of your mons to limit results to}

\item{raid}{is the attacker a raid boss?}

\item{numreturn}{the number of your mons to return, most survivable first}

\item{sourceatk}{specify the attacker's attack IV}

\item{sourcelvl}{specify the attacker's level}

\item{cpmin}{limit results to those over a certain CP}

\item{data}{the name of your database}
}
\description{
Iterates mylist.strong() for each mon in your database
That is, calculates the pokemon who take the least damage from a source
}
