% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.subtract}
\alias{mylist.subtract}
\title{Subtract a pokemon from the database}
\usage{
mylist.subtract(num, data = mymons)
}
\arguments{
\item{num}{The row number in the database}

\item{data}{The database}
}
\description{
Subtract a pokemon from the database
}
