% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{mylist.update}
\alias{mylist.update}
\title{Updates IV stats for one mon in your database}
\usage{
mylist.update(num, Attack, Defense, Stamina, Notes = NA, data = mymons)
}
\arguments{
\item{num}{row number of the mon}

\item{Attack}{new attack stat}

\item{Defense}{new defense stat}

\item{Stamina}{new stamina stat}

\item{Notes}{new note}

\item{data}{the name of your database}
}
\description{
Updates IV stats for one mon in your database
}
