% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{nearmons}
\alias{nearmons}
\title{Checks the pokedex (more or less), shows the names of those around a given name}
\usage{
nearmons(name, range = 3)
}
\arguments{
\item{name}{the name to look up}

\item{range}{the number of mons above and below to display}
}
\description{
Checks the pokedex (more or less), shows the names of those around a given name
}
