% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{paste_from_clipboard}
\alias{paste_from_clipboard}
\title{Paste info from the clipboard}
\usage{
paste_from_clipboard(sep = "\\t", header = T, ...)
}
\arguments{
\item{sep}{The cell separator}

\item{header}{First row is a header}
}
\value{
Dataframe
}
\description{
Paste info from the clipboard
}
