% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{pctcalc}
\alias{pctcalc}
\title{Just calculate your IV percentage (sum of stats / 45)}
\usage{
pctcalc(Attack, Defense, Stamina)
}
\arguments{
\item{Attack}{attack stat}

\item{Defense}{defense stat}

\item{Stamina}{stamina stat}
}
\description{
Just calculate your IV percentage (sum of stats / 45)
}
