% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{possibles}
\alias{possibles}
\title{Take pokemon, CP, HP, dust to level, and optionally descriptors of stats
to find all possible combinations of level and stats.
For atk/def/stam, descriptors are "not"/"pos"/"impr"/"incr", for the summary
they are "not"/"aa"/"caught"/"wonder". (As they were for Team Mystic under the
old assessment procedure. Not so useful anymore!)}
\usage{
possibles(
  mon,
  cp,
  hp,
  dust = NA,
  atk = "all",
  def = "all",
  stam = "all",
  total = "all",
  powered = F,
  evolve = NA
)
}
\arguments{
\item{mon}{The mon in question}

\item{cp}{Its CP}

\item{hp}{Its HP}

\item{dust}{Optionally, the dust it takes to reach its next level}

\item{atk}{How its attack is described}

\item{def}{How its defense is described}

\item{stam}{How its stamina is described}

\item{total}{How its summary is described}

\item{powered}{Has it been powered up since being obtained (can it be a 'half' level)?}

\item{evolve}{Add a different mon to show CP of that other pokemon (at level 40) compared to its max instead.}
}
\description{
Take pokemon, CP, HP, dust to level, and optionally descriptors of stats
to find all possible combinations of level and stats.
For atk/def/stam, descriptors are "not"/"pos"/"impr"/"incr", for the summary
they are "not"/"aa"/"caught"/"wonder". (As they were for Team Mystic under the
old assessment procedure. Not so useful anymore!)
}
