% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{raidbosses.add}
\alias{raidbosses.add}
\title{Update the raidbosses list with a previously unknown raid boss. Add at least name and tier.
Everything but name and tier isn't strictly necessary.}
\usage{
raidbosses.add(
  name,
  Tier,
  CP = 0,
  Attack = 0,
  Defense = 0,
  HP = 0,
  Catch.Rate = NA,
  Active = TRUE,
  ID = NA
)
}
\arguments{
\item{name}{the name of the pokemon type}

\item{CP}{its CP}

\item{Attack}{its attack}

\item{Defense}{its defense}

\item{HP}{its HP}

\item{Catch.Rate}{chance to catch it once the raid's over}

\item{Active}{is the boss now active?}

\item{ID}{its pokedex ID}

\item{tier}{the tier number}
}
\description{
Update the raidbosses list with a previously unknown raid boss. Add at least name and tier.
Everything but name and tier isn't strictly necessary.
}
