% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{raidbosses.get}
\alias{raidbosses.get}
\title{Get list of current raid bosses}
\usage{
raidbosses.get(tier = 1:5)
}
\arguments{
\item{tier}{what tier (1:5) to list}
}
\value{
a slice of raidbosses corresponding to the selected tier(s)
}
\description{
Get list of current raid bosses
}
