% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{raidbosses.set}
\alias{raidbosses.set}
\title{Switch up the list of active raid bosses
Defaults to clearing all tiers!}
\usage{
raidbosses.set(bosslist, tier = 1:5)
}
\arguments{
\item{bosslist}{a list (actually a character vector) of raid bosses to activate}

\item{tier}{resets all members of that raid tier to inactive before making new ones active}
}
\description{
Switch up the list of active raid bosses
Defaults to clearing all tiers!
}
