% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{raidweather}
\alias{raidweather}
\title{Lists all active raid bosses enhanced by current weather.}
\usage{
raidweather(weather = NA, inc.inactive = F)
}
\description{
Lists all active raid bosses enhanced by current weather.
}
