% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{renum}
\alias{renum}
\title{Little helper: Takes the name of a data frame (as a string!) and renumbers it (deletes rownames).
Changes done in place.}
\usage{
renum(dataname)
}
\arguments{
\item{dataname}{any dataframe}
}
\description{
Little helper: Takes the name of a data frame (as a string!) and renumbers it (deletes rownames).
Changes done in place.
}
