% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{strengths}
\alias{strengths}
\title{Shows the damage type advantages of a given type of attack}
\usage{
strengths(types)
}
\arguments{
\item{types}{A vector of types}
}
\value{
A table
}
\description{
Shows the damage type advantages of a given type of attack
}
