% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{strongest}
\alias{strongest}
\title{Lists the known movesets that are strongest against the target, with an
option to choose a particular pokemon as the attack source, though that's
similar to using alldps().}
\usage{
strongest(target, source = NA)
}
\arguments{
\item{target}{The pokemon to attack}

\item{source}{Limit to a particular attacker}
}
\description{
Lists the known movesets that are strongest against the target, with an
option to choose a particular pokemon as the attack source, though that's
similar to using alldps().
}
