% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{totaldps}
\alias{totaldps}
\title{Gives the combined average DPS per second of a pair of moves.}
\usage{
totaldps(move1, move2, source = NA, target = NA, stab1 = TRUE,
  stab2 = TRUE)
}
\arguments{
\item{move1}{The fast move}

\item{move2}{The charge move}

\item{source}{Source pokemon}

\item{target}{Target pokemon}

\item{stab1}{Does the first move get a STAB bonus?}

\item{stab2}{Does the second move get a STAB bonus?}
}
\description{
Gives the combined average DPS per second of a pair of moves.
}
