% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\name{weaknesses}
\alias{weaknesses}
\title{Looks up a mon's weaknesses, compounds them if multiple types, excludes uninteresting (==1) values.}
\usage{
weaknesses(mon, all = FALSE)
}
\arguments{
\item{mon}{The pokemon in question}

\item{all}{Show the regular effectiveness}
}
\description{
Looks up a mon's weaknesses, compounds them if multiple types, excludes uninteresting (==1) values.
}
